refactor: move window tests into window file
The recommended standard for unit tests in Rust is to include the unit tests in the file that they're testing
